moved out event to seperate view
