+++ funnel/views/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import index, login, profile, space, section, usergroup, proposal, commentvote, venue, schedule, session
+from . import index, login, profile, space, section, usergroup, proposal, commentvote, venue, schedule, session, event
